Add Counter.sol and refacto tests
