+++ pyproject.toml
@@ -100,6 +100,7 @@ markers = [
     "REVERT:         Opcode value 0xfd - Halt execution reverting state changes",
     "INVALID:        Opcode Value 0xfe - Designated invalid instruction",
     "SolmateERC20",
+    "SolmateERC721",
     "IntegrationTestContract",
 ]
