+++ tests/solidity_files/IntegrationTestContract.sol
@@ -1,26 +1,40 @@
 // SPDX-License-Identifier: AGPL-3.0-only
 pragma solidity >=0.8.0;
 
+interface ICounter {
+    function count() external view returns (uint256);
+
+    function inc() external;
+
+    function dec() external;
+
+    function reset() external;
+}
+
 /// @notice Contract for integration testing of EVM opcodes.
-/// @author Kakarot9000 
+/// @author Kakarot9000
 /// @dev Do add functions and storage variables for opcodes accordingly.
 contract IntegrationTestContract {
     /*//////////////////////////////////////////////////////////////
                             METADATA STORAGE
     //////////////////////////////////////////////////////////////*/
+    ICounter counter;
 
     /*//////////////////////////////////////////////////////////////
                                CONSTRUCTOR
     //////////////////////////////////////////////////////////////*/
+    constructor(address counterAddress) {
+        counter = ICounter(counterAddress);
+    }
 
     /*//////////////////////////////////////////////////////////////
                             FUNCTIONS FOR OPCODES
     //////////////////////////////////////////////////////////////*/
-    function opcodeAddress()
-        public
-        view
-        returns (address selfAddress) 
-    {
+    function opcodeAddress() public view returns (address selfAddress) {
         return (address(this));
+    }
+
+    function opcodeCall() public view returns (uint256) {
+        return counter.count();
     }
 }
