+++ tests/solidity_files/output/IntegrationTestContract.abi
@@ -1,1 +1,1 @@
-[{"inputs":[],"name":"opcodeAddress","outputs":[{"internalType":"address","name":"selfAddress","type":"address"}],"stateMutability":"view","type":"function"}]
+[{"inputs":[{"internalType":"address","name":"counterAddress","type":"address"}],"stateMutability":"nonpayable","type":"constructor"},{"inputs":[],"name":"opcodeAddress","outputs":[{"internalType":"address","name":"selfAddress","type":"address"}],"stateMutability":"view","type":"function"},{"inputs":[],"name":"opcodeCall","outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"stateMutability":"view","type":"function"}]
