+++ tests/solidity_files/output/IntegrationTestContract.bin
@@ -1,1 +1,1 @@
-608060405234801561001057600080fd5b5060db8061001f6000396000f3fe6080604052348015600f57600080fd5b506004361060285760003560e01c8063240070ab14602d575b600080fd5b60336047565b604051603e9190608c565b60405180910390f35b600030905090565b600073ffffffffffffffffffffffffffffffffffffffff82169050919050565b6000607882604f565b9050919050565b608681606f565b82525050565b6000602082019050609f6000830184607f565b9291505056fea264697066735822122034e2af19c75d919cece5516718009a4167a40ff8e01b16fd258e26be904f3cfd64736f6c63430008110033
+608060405234801561001057600080fd5b50604051610351380380610351833981810160405281019061003291906100db565b806000806101000a81548173ffffffffffffffffffffffffffffffffffffffff021916908373ffffffffffffffffffffffffffffffffffffffff16021790555050610108565b600080fd5b600073ffffffffffffffffffffffffffffffffffffffff82169050919050565b60006100a88261007d565b9050919050565b6100b88161009d565b81146100c357600080fd5b50565b6000815190506100d5816100af565b92915050565b6000602082840312156100f1576100f0610078565b5b60006100ff848285016100c6565b91505092915050565b61023a806101176000396000f3fe608060405234801561001057600080fd5b50600436106100365760003560e01c80632328ea801461003b578063240070ab14610059575b600080fd5b610043610077565b604051610050919061012f565b60405180910390f35b61006161010e565b60405161006e919061018b565b60405180910390f35b60008060009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166306661abd6040518163ffffffff1660e01b8152600401602060405180830381865afa1580156100e5573d6000803e3d6000fd5b505050506040513d601f19601f8201168201806040525081019061010991906101d7565b905090565b600030905090565b6000819050919050565b61012981610116565b82525050565b60006020820190506101446000830184610120565b92915050565b600073ffffffffffffffffffffffffffffffffffffffff82169050919050565b60006101758261014a565b9050919050565b6101858161016a565b82525050565b60006020820190506101a0600083018461017c565b92915050565b600080fd5b6101b481610116565b81146101bf57600080fd5b50565b6000815190506101d1816101ab565b92915050565b6000602082840312156101ed576101ec6101a6565b5b60006101fb848285016101c2565b9150509291505056fea26469706673582212201b72c351785210e4c45a807b7fa8a403b3fb33c9dab414378711cfc0128ae0f364736f6c63430008110033
